Moved the PID class to its own file
